(GH-109) Changed rule from Guideline to Requirment
